Reason for change: 1st tag for 2405 Federated.
Test Procedure: IP Playbacks
Risks: Medium

+++ test/utests/fakes/FakeAampGstPlayer.cpp
@@ -86,7 +86,7 @@ long AAMPGstPlayer::GetDurationMilliseconds(void)
 	return 0;
 }
 
-long AAMPGstPlayer::GetPositionMilliseconds(void)
+long long AAMPGstPlayer::GetPositionMilliseconds(void)
 {
 	return 0;
 }
